backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.3.0
pydantic>=2.0.0,<3.0.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
cerebras-cloud-sdk>=1.0.0
groq>=0.4.1
python-dotenv>=1.0.0
python-dotenv
python-dotenv==1.0.0